refactoring Accounts app
